requirements.txt
Pillow
numpy
scipy
matplotlib
scikit-image
tensorflow